requirements.txt
discord.py==0.16.12
requests==2.19.1
mcstatus==2.2
